unionId问题记录.txt

+++ socialuni-server/socialuni-community-sdk/src/main/java/com/socialuni/social/community/sdk/repository/TalkAdultImgAuditRepository.java
@@ -9,5 +9,5 @@ public interface TalkAdultImgAuditRepository extends JpaRepository<SocialuniTalk
 
     List<SocialuniTalkHasUnderageImgAuditDO> findTop20ByStatusAndDevIdOrderByIdAsc(String status, Integer devId);
 
-    SocialuniTalkHasUnderageImgAuditDO findOneById(Integer id);
+    SocialuniTalkHasUnderageImgAuditDO findOneById(Long id);
 }
